+++ src/app/app.routes.ts
@@ -19,11 +19,15 @@ export const routes: Routes = [
         loadChildren: () => import('./views/dashboard/routes').then((m) => m.routes)
       },
       {
-        path: 'utilsateurs',
+        path: 'map',
         loadChildren: () =>
-          import('./views/utilisateur/app-routing.module').then((m) => m.AppRoutingModule)
+          import('./leaflet-map/leaflet-map-routing.module').then((m) => m.LeafletMapRoutingModule)
       },
-
+      // {
+      //   path: 'map',
+      //   loadComponent: () => import('./leaflet-map/leaflet-map.module').then(m => m.LeafletMapModule),
+      //   // component: LeafletMapComponent
+      // },
       {
         path: 'theme',
         loadChildren: () => import('./views/theme/routes').then((m) => m.routes)
@@ -83,6 +87,7 @@ export const routes: Routes = [
       title: 'Login Page'
     }
   },
+
   {
     path: 'register',
     loadComponent: () => import('./views/pages/register/register.component').then(m => m.RegisterComponent),