+++ src/app/views/base/EtatRobot/EtatRobot.component.ts
@@ -1,0 +1,114 @@
+import { CommonModule } from '@angular/common';
+import { HttpClientModule } from '@angular/common/http';
+import { Component, OnInit } from '@angular/core';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
+// import { DocsExampleComponent } from '@docs-components/public-api';
+import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
+
+import {
+  AlertModule,
+  AlignDirective,
+  AvatarModule,
+  BadgeModule,
+  BorderDirective,
+  ButtonGroupModule,
+  ButtonModule,
+  CardBodyComponent,
+  CardComponent,
+  CardHeaderComponent,
+  CardModule,
+  ColComponent,
+  DropdownModule,
+  FormModule,
+  GridModule,
+  NavModule,
+  PaginationModule,
+  ProgressModule,
+  RowComponent,
+  SharedModule,
+  TableActiveDirective,
+  TableColorDirective,
+  TableDirective,
+  TableModule,
+  TabsModule,
+  TextColorDirective,
+  WidgetModule
+} from '@coreui/angular';
+import { ChartjsModule } from '@coreui/angular-chartjs';
+import { IconModule } from '@coreui/icons-angular';
+import { NgxPaginationModule } from 'ngx-pagination';
+
+type CardColor = {
+  color: string
+  textColor?: string
+}
+
+@Component({
+  selector: 'app-EtatRobot',  // Nom de sélecteur en camelCase
+  templateUrl: './EtatRobot.component.html',  // Correspond au fichier renommé
+  styleUrls: ['./EtatRobot.component.scss'],
+  standalone: true,
+  imports: [
+    RowComponent, 
+    ColComponent, 
+    TextColorDirective, 
+    CardComponent, 
+    CardHeaderComponent, 
+    CardBodyComponent, 
+    // DocsExampleComponent, 
+    TableDirective, 
+    TableColorDirective, 
+    TableActiveDirective, 
+    BorderDirective, 
+    AlignDirective,
+    CommonModule,
+    CardModule,
+    NavModule,
+    IconModule,
+    TabsModule,
+    GridModule,
+    ProgressModule,
+    ReactiveFormsModule,
+    ButtonModule,
+    FormModule,
+    FormsModule,
+    HttpClientModule,
+    ButtonModule,
+    ButtonGroupModule,
+    ChartjsModule,
+    AvatarModule,
+    TableModule,
+    //WidgetsModule,
+    ChartjsModule,
+    CardModule,
+    GridModule,
+    BadgeModule,
+    PaginationModule,
+    NgxPaginationModule,
+    // Ng2SearchPipeModule,
+    WidgetModule,
+    DropdownModule,
+    SharedModule,
+    AlertModule
+  
+  ]
+})
+export class EtatRobotComponent  implements OnInit{
+
+  constructor(private modalService: NgbModal) { }
+
+  ngOnInit(): void {
+
+  }
+
+
+  openLg(content: any) {
+		this.modalService.open(content, { size: 'lg' });
+	}
+	open(content: any) {
+		this.modalService.open(content);
+	}
+}
+
+
+