+++ src/app/views/base/routes.ts
@@ -9,7 +9,7 @@ export const routes: Routes = [
     children: [
       {
         path: '',
-        redirectTo: 'cards',
+        redirectTo: 'EtatRobot',
         pathMatch: 'full'
       },
       {
@@ -27,17 +27,17 @@ export const routes: Routes = [
         }
       },
       {
-        path: 'cards',
-        loadComponent: () => import('./cards/cards.component').then(m => m.CardsComponent),
+        path: 'EtatRobot',
+        loadComponent: () => import('./EtatRobot/EtatRobot.component').then(m => m.EtatRobotComponent),
         data: {
-          title: 'Cards'
+          title: 'EtatRobot'
         }
       },
       {
-        path: 'carousel',
-        loadComponent: () => import('./carousels/carousels.component').then(m => m.CarouselsComponent),
+        path: 'utilisateurs',
+        loadComponent: () => import('./utilisateurs/utilisateurs.component').then(m => m.UtilisateursComponent),
         data: {
-          title: 'Carousel'
+          title: 'utilisateurs'
         }
       },
       {
@@ -62,11 +62,15 @@ export const routes: Routes = [
         }
       },
       {
-        path: 'pagination',
-        loadComponent: () => import('./paginations/paginations.component').then(m => m.PaginationsComponent),
+        path: 'categorie',
+        loadComponent: () => import('./categorie/categorie.component').then(m => m.categoriesComponent),
         data: {
-          title: 'Pagination'
+          title: 'categorie'
         }
+      },
+      {
+        path: 'update-categorie/:id',
+        loadComponent: () => import('./categorie/update-categorie/update-categorie.component').then(m => m.UpdateCategorieComponent),
       },
       {
         path: 'placeholder',
@@ -101,6 +105,13 @@ export const routes: Routes = [
         loadComponent: () => import('./tables/tables.component').then(m => m.TablesComponent),
         data: {
           title: 'Tables'
+        }
+      },
+      {
+        path: 'formation',
+        loadComponent: () => import('./formation/formation.component').then(m => m.FormationComponent),
+        data: {
+          title: 'Formation'
         }
       },
       {