+++ src/app/views/pages/login/login.component.ts
@@ -1,22 +1,25 @@
-import { Component } from '@angular/core';
-import { Router, RouterModule } from '@angular/router'; // Import RouterModule for navigation
-import { FormsModule } from '@angular/forms';
 import { CommonModule } from '@angular/common'; // Import CommonModule for ngStyle and other common directives
-import { IconDirective } from '@coreui/icons-angular';
+import { Component, OnInit } from '@angular/core';
+import { FormsModule } from '@angular/forms';
+import { Router, RouterModule } from '@angular/router'; // Import RouterModule for navigation
 import {
-  ContainerComponent,
-  RowComponent,
-  ColComponent,
-  CardGroupComponent,
-  CardComponent,
+  ButtonDirective,
   CardBodyComponent,
+  CardComponent,
+  CardGroupComponent,
+  ColComponent,
+  ContainerComponent,
+  FormControlDirective,
   FormDirective,
   InputGroupComponent,
   InputGroupTextDirective,
-  FormControlDirective,
-  ButtonDirective,
+  RowComponent,
   TextColorDirective,
 } from '@coreui/angular';
+import { IconDirective } from '@coreui/icons-angular';
+import { Login } from '../../../models/login';
+import { AuthService } from '../../../services/auth.services';
+import { StorageService } from '../../../services/storage.services';
 
 @Component({
   selector: 'app-login',
@@ -42,29 +45,82 @@ import {
     TextColorDirective,
   ],
 })
-export class LoginComponent {
-  // Form data model
-  form = {
-    username: '',
-    password: '',
+export class LoginComponent implements OnInit {
+  type= true;
+  changeType(){
+    this.type = !this.type;
+  }
+  signinData: Login = {usernameOrEmail: '', password: ''};
+  popupResult!: boolean;
+  popupMessage = '';
+
+  showAdminBoard = false;
+  showModeratorBoard = false;
+  showUserBoard = false;
+  username?: string;
+
+  form: any = {
+    username: null,
+    password: null,
   };
 
-  // Inject Router service for navigation
-  constructor(private router: Router) {}
+  isLoggedIn = false;
+  isLoginFailed = false;
+  errorMessage = '';
+  roles: string[] = [];
 
-  // Form submission handler
-  onSubmit() {
-    // Example validation logic
-    const isAuthenticated =
-      this.form.username === 'admin' && this.form.password === 'admin';
+  constructor(
+    private router: Router,
+    private authService: AuthService,
+    private storage: StorageService
+  ) {}
 
-    if (isAuthenticated) {
-      // Navigate to the dashboard on successful login
-      this.router.navigate(['/dashboard']); // Update this route as needed
-    } else {
-      // Handle authentication failure
-      console.error('Authentication failed - check username and password');
-      alert('Authentication failed, please try again!');
+  ngOnInit(): void {
+    this.isLoggedIn = !!this.storage.getToken();
+    if (this.isLoggedIn) {
+      const user = this.storage.getUser();
+      this.roles = user.roles;
+      this.showAdminBoard = this.roles.includes('ROLE_ADMIN');
+      this.showUserBoard = this.roles.includes('ROLE_USER');
+      this.username = user.username;
+    }
+    if (this.storage.getToken()) {
+      this.isLoggedIn = true;
+      this.roles = this.storage.getUser().roles;
     }
   }
+
+
+  onSubmit(): void {
+    const { username, password } = this.form;
+    this.authService.login(username, password).subscribe({
+      next: (data) => {
+        this.storage.saveToken(data.accessToken);
+        this.storage.saveUser(data);
+        this.isLoginFailed = false;
+        this.isLoggedIn = true;
+        this.roles = this.storage.getUser().roles;
+        //this.reloadPage();
+        this.navigate();
+      },
+      error: (err) => {
+        this.errorMessage = err.error.message;
+        this.isLoginFailed = true;
+      },
+    });
+  }
+
+  logout(): void {
+    this.storage.signOut();
+    window.location.reload();
+  }
+
+  navigate() {
+    this.router.navigateByUrl('/dashboard');
+  }
+
+  reloadPage(): void {
+    window.location.reload();
+  }
+  
 }