+++ src/app/views/profil/profil-routing.module.ts
@@ -1,0 +1,20 @@
+import { NgModule } from '@angular/core';
+import { RouterModule, Routes } from '@angular/router';
+import { ProfilComponent } from './profil.component';
+
+const routes: Routes = [
+    {
+      path: '',
+      component: ProfilComponent,
+      data: {
+        title: 'Profile'
+      }
+    }
+  ];
+  
+  @NgModule({
+    imports: [RouterModule.forChild(routes)],
+    exports: [RouterModule]
+  })
+  export class ProfilRoutingModule { }
+